pyproject.toml
[tool.poetry]
name = "reportextractorpy"
version = "0.1.0"
description = ""
authors = ["nicksunderland <nicholas.p.sunderland@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
PyQt5 = "5.15.9"
gatenlp = "1.0.8"
nltk = "3.8.1"
spacy = "3.5.2"
numerizer = "0.2.2"
pyYAML = "6.0"
importlib-resources = "5.12.0"
regex = "2023.3.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
